started comments. renamed phase component to BehaviorContainer. moved numberselection to ./client/src/components.
